python/requirements.txt
pycaw == 20181226
comtypes == 1.1.10
pyserial == 3.5